+++ normalize/lib/utils.dart
@@ -1,0 +1,9 @@
+export 'src/config/normalization_config.dart';
+
+export 'src/utils/add_typename_visitor.dart';
+export 'src/utils/get_operation_definition.dart';
+export 'src/utils/resolve_root_typename.dart';
+export 'src/utils/resolve_data_id.dart';
+export 'src/utils/field_key.dart';
+export 'src/utils/is_dangling_reference.dart';
+export 'src/utils/identify.dart';